requirements.txt
git+https://github.com/metabrainz/brainzutils-python.git@v1.11.0
annoy == 1.16.0
cmake == 3.14.4
Flask-Admin == 1.5.3
Flask-Login == 0.4.1
Flask-SQLAlchemy == 2.3.2
Flask-Testing == 0.7.1
Flask-WTF == 0.14.2
matplotlib == 2.2.4
mock == 2.0.0
MulticoreTSNE == 0.1
musicbrainzngs == 0.6
ndg-httpsclient==0.5.1
numpy == 1.14.3
pandas == 0.24.2
plotly == 4.1.0
psutil == 5.6.3
psycopg2 == 2.7.7
pytz==2018.9
pyyaml == 3.13
rauth == 0.7.3
seaborn == 0.9.0
setproctitle == 1.1.10
six == 1.11.0
sklearn == 0.0
-r docs/requirements.txt
